Remove redundant grammar properties
The `name` property of language grammars is only used to provide a human
readable identifier in GUIs. For "injection-only" grammars such as this,
it's not only redundant, but confusing for users who happen across it in
the grammar-selector.

The `fileTypes` array is also optional: its absence is assumed to mean a
grammar has no associated file extensions. An empty array is redundant.

References: atom/grammar-selector#34